requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
marshmallow==3.19.0
packaging==23.1
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
